Testing...

+++ src/avrweather.h
@@ -111,21 +111,21 @@ decimal24_t;
 #define populate_decimal(i, m)      { .integral = i, .mantissa = m }
 
 typedef struct {
-    float           temperature;
-    float           pressure;
-    float           humidity;
+    uint16_t           temperature;
+    uint16_t           pressure;
+    uint16_t           humidity;
 }
 TPH;
 
 typedef struct {
-    float           avgWindspeed;
-    float           maxWindspeed;
+    uint16_t           avgWindspeed;
+    uint16_t           maxWindspeed;
 }
 WINDSPEED;
 
 typedef struct {
-    float           avgRainfall;
-    float           totalRainfall;
+    uint16_t           avgRainfall;
+    uint16_t           totalRainfall;
 }
 RAINFALL;
